refactor(command): extract palette grouping
